config-schema: split out schema view components
