+++ plugins/config-schema/src/components/SchemaView/SchemaView.tsx
@@ -1,0 +1,62 @@
+/*
+ * Copyright 2021 Spotify AB
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import React from 'react';
+import { ArrayView } from './ArrayView';
+import { MatchView } from './MatchView';
+import { ObjectView } from './ObjectView';
+import { ScalarView } from './ScalarView';
+import { SchemaViewProps } from './types';
+
+export function SchemaView(props: SchemaViewProps) {
+  const { schema } = props;
+  if (schema.anyOf) {
+    return (
+      <MatchView
+        {...props}
+        schema={schema.anyOf}
+        label="Any of the following"
+      />
+    );
+  }
+  if (schema.oneOf) {
+    return (
+      <MatchView
+        {...props}
+        schema={schema.oneOf}
+        label="One of the following"
+      />
+    );
+  }
+  if (schema.allOf) {
+    return (
+      <MatchView
+        {...props}
+        schema={schema.allOf}
+        label="All of the following"
+      />
+    );
+  }
+  switch (schema.type) {
+    case 'array':
+      return <ArrayView {...props} />;
+    case 'object':
+    case undefined:
+      return <ObjectView {...props} />;
+    default:
+      return <ScalarView {...props} />;
+  }
+}
